Track basic block successors and explode call target loop

+++ language/src/main/java/au/mrcat/rivet/nodes/RivetDivergentNode.java
@@ -3,7 +3,7 @@ package au.mrcat.rivet.nodes;
 import com.oracle.truffle.api.frame.VirtualFrame;
 
 public abstract class RivetDivergentNode extends RivetNode {
-    public abstract long executeDivergent(VirtualFrame frame);
+    public abstract int executeDivergent(VirtualFrame frame);
     public abstract Long[] callTargetContinuations();
 
     @Override
